requirements.txt
python-dotenv
streamlit 
google-generativeai 
langchain 
google-auth 
unstructured
langchain_google_genai
langchain-community
faiss-cpu
streamlit-extras
streamlit-card 
streamlit-lottie 
streamlit-timeline
requests 
fpdf2
